refactor: extract eusart from init module
