+++ firmware/development/kampela_experiments_efm32pg23/app3/src/peripherals/mod.rs
@@ -1,5 +1,6 @@
 pub mod adc;
 pub mod cmu;
+pub mod eusart;
 pub mod gpio_pins;
 pub mod se_command;
 pub mod usart;
